refactor(ms-bank-account): change all request and response types from bigdecimal to double and fix the response date formatting

+++ bank-account/src/main/java/com/dac/bank_account/command/dto/request/AccountRequestDTO.java
@@ -11,7 +11,7 @@ public record AccountRequestDTO(
 
         @NotNull
         @Positive
-        BigDecimal salary,
+        Double salary,
 
         @NotNull
         Long managerId
